Src folder structure refactor
