+++ app/test/config.test.js
@@ -1,5 +1,5 @@
 const chai = require('chai');
-const configLoader = require('../src/configLoader');
+const configLoader = require('../src/lib/configLoader');
 
 const should = chai.should();
